Get the most accurate information you can get from NONMEM

Get the most accurate information you can get from NONMEM

Usage

nminfo(
  file,
  mod = ".mod",
  xml = ".xml",
  ext = ".ext",
  cov = ".cov",
  phi = ".phi",
  lst = ".lst",
  useXml = TRUE,
  useExt = TRUE,
  useCov = TRUE,
  usePhi = TRUE,
  useLst = TRUE,
  strictLst = FALSE,
  verbose = FALSE
)

Arguments

file

nonmem file, like control stream, phi. This function will remove the extension to try to get the right information. It preferentially selects the most accurate estimates from the file.

mod

the NONMEM output extension, defaults to .mod

xml

the NONMEM xml file extension , defaults to .xml

ext

the NONMEM ext file extension, defaults to .ext

cov

the NONMEM covariance file extension, defaults to .cov

phi

the NONMEM eta/phi file extension, defaults to .phi

lst

the NONMEM output extension, defaults to .lst

useXml

if present, use the NONMEM xml file to import much of the NONMEM information

useExt

if present, use the NONMEM ext file to extract parameter estimates (default TRUE), otherwise defaults to parameter estimates extracted in the NONMEM output

useCov

if present, use the NONMEM cov file to import the covariance, otherwise import the covariance with list file

usePhi

if present, use the NONMEM phi file to extract etas (default TRUE), otherwise defaults to etas in the tables (if present)

useLst

if present, use the NONMEM lst file to extract NONMEM information

strictLst

The list parsing needs to be correct for a successful load (default FALSE).

verbose

this is a flag to be more verbose when reading information in, by default this is FALSE

Value

list of NONMEM information

Author

Matthew L. Fidler

Examples

nminfo(system.file("mods/cpt/runODE032.res", package="nonmem2rx"))
#> $theta
#>    theta1    theta2    theta3    theta4    theta5 
#> 1.3703404 4.1981491 1.3800349 3.8765734 0.1964461 
#> 
#> $omega
#>           eta1       eta2      eta3       eta4
#> eta1 0.1012514 0.00000000 0.0000000 0.00000000
#> eta2 0.0000000 0.09938724 0.0000000 0.00000000
#> eta3 0.0000000 0.00000000 0.1013027 0.00000000
#> eta4 0.0000000 0.00000000 0.0000000 0.07304975
#> 
#> $cov
#>                  theta1        theta2        theta3        theta4        theta5
#> theta1     8.876810e-04 -1.055098e-04  1.844162e-04 -1.202337e-04  5.278300e-08
#> theta2    -1.055098e-04  8.714095e-04 -1.061946e-04 -5.066632e-05 -1.565618e-05
#> theta3     1.844162e-04 -1.061946e-04  2.993363e-03  1.652516e-04  5.993313e-06
#> theta4    -1.202337e-04 -5.066632e-05  1.652516e-04  1.213465e-03 -2.539912e-05
#> theta5     5.278300e-08 -1.565618e-05  5.993313e-06 -2.539912e-05  9.942182e-06
#> eta1      -4.712728e-05  4.696667e-05 -3.642709e-05  2.547962e-05 -8.168847e-06
#> omega.1.2  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> omega.1.3  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> omega.1.4  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> eta2      -7.371560e-05  2.566338e-05 -8.083493e-05  1.369999e-05 -4.365635e-06
#> omega.2.3  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> omega.2.4  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> eta3       6.633832e-05 -8.190016e-05  5.489848e-04  1.683555e-04  1.591222e-06
#> omega.3.4  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> eta4      -9.496613e-06  1.101079e-04 -3.065372e-04 -9.128974e-05  3.187703e-06
#> eps1       0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>                    eta1 omega.1.2 omega.1.3 omega.1.4          eta2 omega.2.3
#> theta1    -4.712728e-05         0         0         0 -7.371560e-05         0
#> theta2     4.696667e-05         0         0         0  2.566338e-05         0
#> theta3    -3.642709e-05         0         0         0 -8.083493e-05         0
#> theta4     2.547962e-05         0         0         0  1.369999e-05         0
#> theta5    -8.168847e-06         0         0         0 -4.365635e-06         0
#> eta1       1.692964e-04         0         0         0  8.751806e-06         0
#> omega.1.2  0.000000e+00         0         0         0  0.000000e+00         0
#> omega.1.3  0.000000e+00         0         0         0  0.000000e+00         0
#> omega.1.4  0.000000e+00         0         0         0  0.000000e+00         0
#> eta2       8.751806e-06         0         0         0  1.512503e-04         0
#> omega.2.3  0.000000e+00         0         0         0  0.000000e+00         0
#> omega.2.4  0.000000e+00         0         0         0  0.000000e+00         0
#> eta3       3.487139e-05         0         0         0  4.315929e-07         0
#> omega.3.4  0.000000e+00         0         0         0  0.000000e+00         0
#> eta4       1.366281e-05         0         0         0 -1.950959e-05         0
#> eps1       0.000000e+00         0         0         0  0.000000e+00         0
#>           omega.2.4          eta3 omega.3.4          eta4 eps1
#> theta1            0  6.633832e-05         0 -9.496613e-06    0
#> theta2            0 -8.190016e-05         0  1.101079e-04    0
#> theta3            0  5.489848e-04         0 -3.065372e-04    0
#> theta4            0  1.683555e-04         0 -9.128974e-05    0
#> theta5            0  1.591222e-06         0  3.187703e-06    0
#> eta1              0  3.487139e-05         0  1.366281e-05    0
#> omega.1.2         0  0.000000e+00         0  0.000000e+00    0
#> omega.1.3         0  0.000000e+00         0  0.000000e+00    0
#> omega.1.4         0  0.000000e+00         0  0.000000e+00    0
#> eta2              0  4.315929e-07         0 -1.950959e-05    0
#> omega.2.3         0  0.000000e+00         0  0.000000e+00    0
#> omega.2.4         0  0.000000e+00         0  0.000000e+00    0
#> eta3              0  9.590290e-04         0 -1.297699e-04    0
#> omega.3.4         0  0.000000e+00         0  0.000000e+00    0
#> eta4              0 -1.297699e-04         0  5.101895e-04    0
#> eps1              0  0.000000e+00         0  0.000000e+00    0
#> 
#> $objf
#> [1] 20167.64
#> 
#> $nobs
#> [1] 2280
#> 
#> $nsub
#> [1] 120
#> 
#> $nmtran
#> [1] "\n\n WARNINGS AND ERRORS (IF ANY) FOR PROBLEM    1\n\n (WARNING  2) NM-TRAN INFERS THAT THE DATA ARE POPULATION.\n"
#> 
#> $termInfo
#> [1] "\n0MINIMIZATION SUCCESSFUL\n NO. OF FUNCTION EVALUATIONS USED:      320\n NO. OF SIG. DIGITS IN FINAL EST.:  2.5\n"
#> 
#> $nonmem
#> [1] "7.4.3"
#> 
#> $time
#> [1] 100.95
#> 
#> $tere
#> [1] " Elapsed estimation  time in seconds:    71.95\n Elapsed covariance  time in seconds:    28.38\n Elapsed postprocess time in seconds:     0.43"
#> 
#> $control
#>  [1] ""                                                                                   
#>  [2] "$PROB    BOLUS_2CPT_CLV1QV2 SINGLE DOSE FOCEI (120 Ind/2280 Obs) runODE032"         
#>  [3] "$INPUT   ID TIME DV LNDV MDV AMT EVID DOSE V1I CLI QI V2I SSX IIX SD CMT"           
#>  [4] "$DATA    BOLUS_2CPT.csv IGNORE=@ IGNORE (SD.EQ.0)"                                  
#>  [5] "$SUBR    ADVAN13 TOL=6"                                                             
#>  [6] "$MODEL"                                                                             
#>  [7] "         COMP=(CENTRAL,DEFOBS,DEFDOSE)"                                             
#>  [8] "         COMP=(PERI)"                                                               
#>  [9] "$PK"                                                                                
#> [10] "         CL=EXP(THETA(1)+ETA(1))"                                                   
#> [11] "         V=EXP(THETA(2)+ETA(2))"                                                    
#> [12] "         Q=EXP(THETA(3)+ETA(3))"                                                    
#> [13] "         V2=EXP(THETA(4)+ETA(4))"                                                   
#> [14] "         V1=V"                                                                      
#> [15] "         S1=V"                                                                      
#> [16] "\t\t K21=Q/V2"                                                                      
#> [17] "\t\t K12=Q/V"                                                                       
#> [18] "$DES"                                                                               
#> [19] "         DADT(1)= K21*A(2)-K12*A(1)-CL*A(1)/V1"                                     
#> [20] "         DADT(2)=-K21*A(2)+K12*A(1)  \t\t"                                          
#> [21] "$ERROR"                                                                             
#> [22] "         IPRED = F"                                                                 
#> [23] "         RESCV = THETA(5)"                                                          
#> [24] "         W     = IPRED*RESCV"                                                       
#> [25] "         IRES  = DV-IPRED"                                                          
#> [26] "         IWRES = IRES/W"                                                            
#> [27] "         Y     = IPRED+W*EPS(1)"                                                    
#> [28] "$THETA   1.6       ;log Cl"                                                         
#> [29] "$THETA   4.5       ;log Vc"                                                         
#> [30] "$THETA   1.6       ;log Q"                                                          
#> [31] "$THETA   4         ;log Vp"                                                         
#> [32] "$THETA   (0,0.3,1) ;RSV"                                                            
#> [33] "$OMEGA   0.15 0.15 0.15 0.15"                                                       
#> [34] "$SIGMA   1 FIX"                                                                     
#> [35] "$EST     NSIG=2 SIGL=6 PRINT=5 MAX=9999 NOABORT POSTHOC METHOD=COND INTER NOOBT"    
#> [36] "$COV"                                                                               
#> [37] "$TABLE  ID TIME LNDV MDV AMT EVID DOSE V1I CLI QI V2I CL V Q V2 ETA1 ETA2 ETA3 ETA4"
#> [38] "        IPRED IRES IWRES CWRESI"                                                    
#> [39] "        ONEHEADER NOPRINT FILE=runODE032.csv"                                       
#> 
#> $eta
#>      ID        eta1        eta2        eta3         eta4
#> 1     1 -0.14424000  0.37464400  0.06501120  0.240662000
#> 2     2  0.56765200 -0.17515700  0.35129700  0.068655100
#> 3     3  0.47739800 -0.05753220 -0.07838230 -0.029594800
#> 4     4 -0.59588800  0.40511500  0.06595780 -0.104262000
#> 5     5 -0.32363600  0.27545000  0.02914670  0.251918000
#> 6     6  0.23277900  0.16120000 -0.00238193  0.064909600
#> 7     7  0.60699200  0.01759660  0.11880500 -0.028699900
#> 8     8  0.31283200 -0.53217800 -0.06375310 -0.221174000
#> 9     9  0.29495300  0.05832140  0.10949400  0.231407000
#> 10   10  0.14195300 -0.24786400 -0.17256300 -0.254001000
#> 11   11 -0.27053800 -0.23474000 -0.15745700  0.170349000
#> 12   12 -0.42602800  0.42758100  0.07017320 -0.075792800
#> 13   13  0.08017620  0.75176000 -0.00136251  0.045172700
#> 14   14 -0.12636300 -0.10582300 -0.08585340  0.020315600
#> 15   15  0.38692100  0.15867700 -0.00800554  0.108208000
#> 16   16  0.27036200  0.18321400 -0.00419876  0.374298000
#> 17   17  0.23923600 -0.32492200 -0.40412600  0.249456000
#> 18   18 -0.00269962  0.15345600  0.08357370  0.229613000
#> 19   19  0.13841300 -0.54122300 -0.03773240  0.296767000
#> 20   20 -0.46801300 -0.20000400  0.22992100  0.519881000
#> 21   21  0.28046300  0.19223800 -0.24784800 -0.053553600
#> 22   22 -0.00235221  0.10750900 -0.01570640 -0.170039000
#> 23   23  0.43714700 -0.09834960  0.20282200 -0.291774000
#> 24   24 -0.22670500  0.10172000 -0.06534150  0.019352400
#> 25   25  0.33364600  0.05095240  0.12533900 -0.284328000
#> 26   26 -0.15186700 -0.26233000  0.02811900 -0.065962400
#> 27   27 -0.28479700 -0.07478660 -0.02914140 -0.074459200
#> 28   28  0.04748590  0.17603800  0.11330300  0.116149000
#> 29   29  0.11232300 -0.42770200 -0.12393100  0.112370000
#> 30   30 -0.10120700  0.24156700  0.18911800  0.022527100
#> 31   31  0.26038200 -0.31787200 -0.46396000 -0.057012300
#> 32   32  0.02265670 -0.12159100  0.09948040 -0.044713800
#> 33   33 -0.31068700 -0.50537800  0.11857700 -0.072940000
#> 34   34  0.15974200 -0.00581950 -0.14293700  0.026960100
#> 35   35  0.05689450 -0.18707700  0.13855200 -0.038802000
#> 36   36 -0.04149100  0.01467420  0.24871300  0.235971000
#> 37   37 -0.28270600  0.18864700 -0.14675500  0.002239020
#> 38   38 -0.39544100 -0.17566300  0.03086270  0.160316000
#> 39   39 -0.44801000 -0.23339500 -0.05457800  0.034569300
#> 40   40 -0.40352400  0.71639900  0.03388700 -0.050975100
#> 41   41 -0.09660390  0.37159800 -0.14032900 -0.163309000
#> 42   42  0.40655200 -0.19898700  0.09422060  0.014364700
#> 43   43  0.06365490 -0.23971500  0.18403100  0.024279300
#> 44   44  0.73216500  0.15155300  0.08640920 -0.070061900
#> 45   45  0.69482300 -0.33599300 -0.45285800  0.327398000
#> 46   46 -0.15903400 -0.49138700  0.08631480 -0.326643000
#> 47   47 -0.32692500  0.52906500  0.15525900 -0.002167200
#> 48   48  0.34205600  0.43590800  0.10205300  0.117408000
#> 49   49  0.15624900  0.12570800 -0.18659000  0.162904000
#> 50   50  0.20167200 -0.27862400 -0.10363700 -0.611866000
#> 51   51  0.28569900 -0.05298130 -0.31277400  0.482237000
#> 52   52 -0.17925200  0.23186600  0.13349000  0.110341000
#> 53   53 -0.20669600 -0.28473000  0.14233500  0.082932900
#> 54   54 -0.14132400  0.37939300 -0.05261020  0.029393800
#> 55   55 -0.11181300 -0.32678600  0.16299000 -0.070609600
#> 56   56 -0.63672900  0.58485000  0.10498900  0.013498600
#> 57   57 -0.21507100 -0.22916600 -0.28756200  0.178900000
#> 58   58  0.35806700 -0.05416930  0.40838100 -0.276384000
#> 59   59  0.05910530 -0.44828200 -0.12018000  0.218146000
#> 60   60  0.10759500 -0.04209500  0.29418100 -0.106099000
#> 61   61  0.37837000 -0.07172710 -0.08466650  0.199045000
#> 62   62 -0.27696000 -0.14019400 -0.10502100  0.037360700
#> 63   63 -0.72832600  0.33906900  0.15471800 -0.157963000
#> 64   64  0.38560500 -0.12710600  0.12557100 -0.306528000
#> 65   65  0.16094300 -0.16399700 -0.15446100  0.044445100
#> 66   66  0.69667500  0.34464000  0.13533300 -0.455244000
#> 67   67  0.44275200  0.12776100  0.36124600 -0.412007000
#> 68   68 -0.19654600  0.11443100  0.10645000 -0.590975000
#> 69   69 -0.05818240 -0.08413010 -0.05650970  0.263846000
#> 70   70  0.14952100  0.39368100 -0.04981730  0.191504000
#> 71   71 -0.33740700  0.10570800  0.19619900  0.155928000
#> 72   72 -0.69872100 -0.51533700  0.04396320  0.000846648
#> 73   73  0.01076690  0.61540700  0.09002220  0.178556000
#> 74   74  0.20712900 -0.21042600 -0.05817590  0.137012000
#> 75   75  0.36452900  0.00945768  0.01538820 -0.102007000
#> 76   76 -0.06147530 -0.01864860  0.03956850  0.025404600
#> 77   77  0.14755500 -0.33821000  0.08060850 -0.104637000
#> 78   78  0.05043080 -0.14039200 -0.06552990  0.037271100
#> 79   79 -0.13078600 -0.28282800  0.17856900 -0.255775000
#> 80   80 -0.04714240  0.13484500  0.18002800  0.108969000
#> 81   81  0.03155160 -0.21060600  0.14035000  0.089137500
#> 82   82  0.37108300 -0.31208900 -0.23774900  0.026305200
#> 83   83  0.25148300  0.61218800  0.10758100 -0.176849000
#> 84   84 -0.18963100  0.04943880 -0.04214050  0.102070000
#> 85   85  0.33298700  0.27384800  0.16310000  0.033666000
#> 86   86 -0.23812200 -0.16038400 -0.08634300 -0.058731500
#> 87   87  0.38391600  0.04119620 -0.27643800  0.147468000
#> 88   88 -0.10298800  0.19976700 -0.15661500 -0.200062000
#> 89   89 -0.15166600  0.41495400  0.09549760 -0.065032300
#> 90   90  0.01407440  0.29514700  0.19409100  0.135821000
#> 91   91 -0.41928800 -0.60078300 -0.32863400  0.113174000
#> 92   92 -0.06073560 -0.19536500  0.16977300 -0.191269000
#> 93   93 -0.08957630  0.22526300  0.09766250  0.051171900
#> 94   94 -0.59527900  0.44926000  0.09443550 -0.091018700
#> 95   95 -0.52511800  0.04418750 -0.49548800 -0.187676000
#> 96   96  0.00218837 -0.08893150  0.05216140 -0.087904800
#> 97   97 -0.01221490 -0.62926800 -0.21896700  0.167880000
#> 98   98 -0.06611430 -0.48380600  0.25682800  0.089878300
#> 99   99  0.28847000 -0.12401300  0.31547400 -0.042283500
#> 100 100  0.16767600 -0.18271000  0.12634200 -0.288271000
#> 101 101  0.13754700 -0.01918670  0.28368400  0.137620000
#> 102 102  0.27452300  0.47222900  0.07034760  0.106941000
#> 103 103 -0.11568100  0.11760300  0.18951200 -0.043268100
#> 104 104  0.06801260  0.05801770 -0.04669300 -0.144975000
#> 105 105 -0.16200900  0.16523000  0.48384200  0.205917000
#> 106 106 -0.10389900  0.39249200 -0.01162800  0.112158000
#> 107 107 -0.46541000 -0.16248700 -0.22631000  0.026516500
#> 108 108 -0.56156500 -0.20425100  0.08323350 -0.067884700
#> 109 109 -0.02578220  0.29599700 -0.03974570  0.154768000
#> 110 110  0.05554480 -0.76315700 -0.49552200  0.011846900
#> 111 111 -0.50922900  0.07079820 -0.10147900 -0.116094000
#> 112 112 -0.19079300  0.30565000 -0.02401360 -0.072402000
#> 113 113 -0.03288730 -0.26397800  0.41920200 -0.419513000
#> 114 114  0.17266700 -0.21803900 -0.23855700 -0.024368500
#> 115 115  0.07875010  0.12835600  0.11664400 -0.218375000
#> 116 116 -0.32527700  0.16564300 -0.00834690  0.059186700
#> 117 117 -0.05887220 -0.37614800  0.09676570 -0.281402000
#> 118 118 -0.13482900  0.25191100 -0.22708900 -0.139859000
#> 119 119  0.10252400 -0.40742900 -0.28432200  0.146728000
#> 120 120  0.55679800 -0.17809900 -0.18922100 -0.178665000
#> 
#> $uses
#> [1] "xml" "ext" "phi" "lst"
#> 
#> $thetaSource
#> [1] "xml"
#> 
#> $omegaSource
#> [1] "xml"
#> 
#> $covSource
#> [1] "xml"
#> 
#> $objfSource
#> [1] "xml"
#> 
#> $sigma
#>      eps1
#> eps1    1
#> 
#> $sigmaSource
#> [1] "ext"
#>