Convert a NONMEM source file to a rxode model (nlmixr2-syle)

Convert a NONMEM source file to a rxode model (nlmixr2-syle)

nonmem2rx(
  file,
  tolowerLhs = TRUE,
  thetaNames = TRUE,
  etaNames = TRUE,
  cmtNames = TRUE,
  updateFinal = TRUE,
  determineError = TRUE,
  validate = TRUE,
  lst = ".lst",
  ext = ".ext"
)

Arguments

file

NONMEM run file

tolowerLhs

Boolean to change the lhs to lower case (default: `TRUE`)

thetaNames

this could be a boolean indicating that the theta names should be changed to the comment-labeled names (default: `TRUE`). This could also be a character vector of the theta names (in order) to be replaced.

etaNames

this could be a boolean indicating that the eta names should be changed to the comment-labeled names (default: `TRUE`). This could also be a character vector of the theta names (in order) to be replaced.

cmtNames

this could be a boolean indicating that the compartment names should be changed to the named compartments in the `$MODEL` by `COMP = (name)` (default: `TRUE`). This could also be a character vector of the compartment names (in order) to be replaced.

updateFinal

Update the parsed model with the model estimates from the `.lst` output file.

determineError

Boolean to try to determine the `nlmixr2`-style residual error model (like `ipred ~ add(add.sd)`), otherwise endpoints are not defined in the `rxode2`/`nlmixr2` model (default: `TRUE`)

validate

Boolean that this tool will attempt to "validate" the model by solving the derived model under pred conditions (etas are zero and eps values are zero)

lst

the NONMEM output extension, defaults to `.lst`

ext

the NONMEM ext file extension, defaults to `.ext`

Value

rxode2 function

Examples

nonmem2rx(system.file("run001.mod", package="nonmem2rx"))
#> Warning: multiple $PROBLEM statements; only use first $PROBLEM for translation
#> Reading data using fread
#> Adding table numbers to data
#> getting rid of non-data rows
#> ℹ change initial estimate of `theta1` to `26.2909`
#> ℹ change initial estimate of `theta2` to `1.34809`
#> ℹ change initial estimate of `theta3` to `4.20364`
#> ℹ change initial estimate of `theta4` to `0.207958`
#> ℹ change initial estimate of `theta5` to `0.20461`
#> ℹ change initial estimate of `theta6` to `0.0105527`
#> ℹ change initial estimate of `theta7` to `0.00717161`
#> ℹ change initial estimate of `eta1` to `0.0729525`
#> ℹ change initial estimate of `eta2` to `0.0380192`
#> ℹ change initial estimate of `eta3` to `1.90699`
#> ℹ done
#> ℹ changing most variables to lower case
#> ℹ done
#> ℹ replace theta names
#> ℹ done
#> ℹ replace eta names
#> ℹ done
#>  ── rxode2-based solved PK 1-compartment model with first-order absorption ────── 
#>  ── Initalization: ──  
#> Fixed Effects ($theta): 
#>      t.TVCL       t.TVV        TVKA         LAG       Prop.      t.Add. 
#> 26.29090000  1.34809000  4.20364000  0.20795800  0.20461000  0.01055270 
#>      t.CRCL 
#>  0.00717161 
#> 
#> Omega ($omega): 
#>           IIV.CL     IIV.V  IIV.KA
#> IIV.CL 0.0729525 0.0000000 0.00000
#> IIV.V  0.0000000 0.0380192 0.00000
#> IIV.KA 0.0000000 0.0000000 1.90699
#>  ── Model (Normalized Syntax): ── 
#> function() {
#>     description <- c(";; 1. Based on:     000", ";; 2. Description:       ", 
#>         ";;    NONMEM PK example for xpose", "    Parameter estimation", 
#>         "    Model simulations")
#>     sigma <- structure(1, dim = c(1L, 1L), dimnames = list("eps1", 
#>         "eps1"))
#>     ini({
#>         t.TVCL <- c(0, 26.2909)
#>         label("TVCL")
#>         t.TVV <- c(0, 1.34809)
#>         label("TVV")
#>         TVKA <- c(0, 4.20364)
#>         label("TVKA")
#>         LAG <- c(0, 0.207958)
#>         label("LAG")
#>         Prop. <- c(0, 0.20461)
#>         label("Prop.Err")
#>         t.Add. <- c(0, 0.0105527)
#>         label("Add.Err")
#>         t.CRCL <- c(0, 0.00717161, 0.02941)
#>         label("CRCL onCL")
#>         IIV.CL ~ 0.0729525
#>         IIV.V ~ 0.0380192
#>         IIV.KA ~ 1.90699
#>     })
#>     model({
#>         tvcl <- t.TVCL * (1 + t.CRCL * (CLCR - 65))
#>         tvv <- t.TVV * WT
#>         cl <- tvcl * exp(IIV.CL)
#>         v <- tvv * exp(IIV.V)
#>         ka <- TVKA * exp(IIV.KA)
#>         alag(depot) <- LAG
#>         k <- cl/v
#>         rxlincmt1 <- linCmt()
#>         f <- rxlincmt1
#>         a1 <- depot
#>         a2 <- central
#>         ipred <- a2/v
#>         ipred ~ add(t.Add.) + prop(Prop.) + combined1()
#>     })
#> }

nonmem2rx(system.file("mods/cpt/runODE032.ctl", package="nonmem2rx"), lst=".res")
#> Warning: $SUBROUTINES TOL=# ignored
#> Reading data using fread
#> Adding table numbers to data
#> getting rid of non-data rows
#> ℹ change initial estimate of `theta1` to `1.37034`
#> ℹ change initial estimate of `theta2` to `4.19815`
#> ℹ change initial estimate of `theta3` to `1.38003`
#> ℹ change initial estimate of `theta4` to `3.87657`
#> ℹ change initial estimate of `theta5` to `0.196446`
#> ℹ change initial estimate of `eta1` to `0.101251`
#> ℹ change initial estimate of `eta2` to `0.0993872`
#> ℹ change initial estimate of `eta3` to `0.101303`
#> ℹ change initial estimate of `eta4` to `0.0730498`
#> Reading /home/runner/work/_temp/Library/nonmem2rx/mods/cpt/runODE032.cov
#> ℹ read in nonmem input data (for model validation): /home/runner/work/_temp/Library/nonmem2rx/mods/cpt/Bolus_2CPT.csv
#> ℹ ignoring lines that begin with a letter (IGNORE=@)'
#> ℹ applying names specified by $INPUT
#> ℹ subsetting accept/ignore filters code: .data[-which((.data$SD == 0)),]
#> ℹ done
#> ℹ read in nonmem IPRED data (for model validation): /home/runner/work/_temp/Library/nonmem2rx/mods/cpt/runODE032.csv
#> Reading data using fread
#> Adding table numbers to data
#> getting rid of non-data rows
#> ℹ done
#> ℹ read in nonmem ETA data (for model validation): /home/runner/work/_temp/Library/nonmem2rx/mods/cpt/runODE032.csv
#> Reading data using fread
#> Adding table numbers to data
#> getting rid of non-data rows
#> ℹ done
#> ℹ changing most variables to lower case
#> ℹ done
#> ℹ replace theta names
#> Warning: there are duplicate theta names, not renaming duplicate parameters
#> ℹ done
#> ℹ replace eta names
#> Warning: there are duplicate eta names, not renaming duplicate parameters
#> ℹ done (no labels)
#> ℹ renaming compartments
#> ℹ done
#>  
#>  
#>  ── rxode2-based free-form 2-cmt ODE model ────────────────────────────────────── 
#>  ── Initalization: ──  
#> Fixed Effects ($theta): 
#>   theta1   theta2   theta3   theta4      RSV 
#> 1.370340 4.198150 1.380030 3.876570 0.196446 
#> 
#> Omega ($omega): 
#>          eta1      eta2     eta3      eta4
#> eta1 0.101251 0.0000000 0.000000 0.0000000
#> eta2 0.000000 0.0993872 0.000000 0.0000000
#> eta3 0.000000 0.0000000 0.101303 0.0000000
#> eta4 0.000000 0.0000000 0.000000 0.0730498
#> 
#> States ($state or $stateDf): 
#>   Compartment Number Compartment Name
#> 1                  1          CENTRAL
#> 2                  2             PERI
#>  ── μ-referencing ($muRefTable): ──  
#>    theta  eta level
#> 1 theta1 eta1    id
#> 2 theta2 eta2    id
#> 3 theta3 eta3    id
#> 4 theta4 eta4    id
#> 
#>  ── Model (Normalized Syntax): ── 
#> function() {
#>     description <- "    BOLUS_2CPT_CLV1QV2 SINGLE DOSE FOCEI (120 Ind/2280 Obs) runODE032"
#>     sigma <- structure(1, dim = c(1L, 1L), dimnames = list("eps1", 
#>         "eps1"))
#>     validation <- c("IPRED relative difference compared to Nonmem IPRED: 0%; 95% percentile: (0%,0%); rtol=7e-06", 
#>         "IPRED absolute difference compared to Nonmem IPRED: atol=0.0018; 95% percentile: (2.41e-05, 0.0421)", 
#>         "PRED relative difference compared to Nonmem PRED: 0%; 95% percentile: (0%,0%); rtol=6.78e-06", 
#>         "PRED absolute difference compared to Nonmem PRED: atol=6.78e-06; 95% percentile: (7.31e-07,0.00322)")
#>     ini({
#>         theta1 <- 1.37034
#>         label("logCl")
#>         theta2 <- 4.19815
#>         label("logVc")
#>         theta3 <- 1.38003
#>         label("logQ")
#>         theta4 <- 3.87657
#>         label("logVp")
#>         RSV <- c(0, 0.196446, 1)
#>         eta1 ~ 0.101251
#>         eta2 ~ 0.0993872
#>         eta3 ~ 0.101303
#>         eta4 ~ 0.0730498
#>     })
#>     model({
#>         cmt(CENTRAL)
#>         cmt(PERI)
#>         cl <- exp(theta1 + eta1)
#>         v <- exp(theta2 + eta2)
#>         q <- exp(theta3 + eta3)
#>         v2 <- exp(theta4 + eta4)
#>         v1 <- v
#>         scale1 <- v
#>         k21 <- q/v2
#>         k12 <- q/v
#>         d/dt(CENTRAL) <- k21 * PERI - k12 * CENTRAL - cl * CENTRAL/v1
#>         d/dt(PERI) <- -k21 * PERI + k12 * CENTRAL
#>         f <- CENTRAL/scale1
#>         ipred <- f
#>         rescv <- RSV
#>         ipred ~ prop(RSV)
#>     })
#> }